ConfigurableArea Test Tidying
